backend/requirements.txt
requests
python-dotenv
langdetect
streamlit
pytest
playwright
pytest-playwright
fastapi
uvicorn
pydantic
sqlmodel
